+++ doublex_expects/__init__.py
@@ -2,10 +2,10 @@
 
 from .matchers import (
     have_been_called, have_been_called_with, have_been_satisfied,
-    anything, any_arg
+    have_been_satisfied_in_any_order, anything, any_arg
 )
 
 __all__ = [
     'have_been_called', 'have_been_called_with', 'have_been_satisfied',
-    'anything', 'any_arg'
+    'have_been_satisfied_in_any_order', 'anything', 'any_arg'
 ]